IOS-2802 Refactor logger
