+++ Tangem/App/Services/Common/AppLog.swift
@@ -1,0 +1,12 @@
+//
+//  AppLog.swift
+//  Tangem
+//
+//  Created by Alexander Osokin on 30.12.2022.
+//  Copyright © 2022 Tangem AG. All rights reserved.
+//
+
+import Foundation
+import TangemSdk
+
+typealias AppLog = Log
