--- Tangem/App/Services/Logger/LoggerProvider.swift
@@ -1,13 +1,0 @@
-//
-//  LoggerProvider.swift
-//  Tangem
-//
-//  Created by Alexander Osokin on 04.05.2022.
-//  Copyright © 2022 Tangem AG. All rights reserved.
-//
-
-import Foundation
-
-struct LoggerProvider: LoggerProviding {
-    var logger: Logger = .init()
-}
