+++ Tangem/App/Services/TangemSdk/TangemSdkConfigFactory.swift
@@ -10,13 +10,11 @@ import Foundation
 import TangemSdk
 
 struct TangemSdkConfigFactory {
-    @Injected(\.loggerProvider) var loggerProvider: LoggerProviding
-
     func makeDefaultConfig() -> Config {
         var config = Config()
         config.filter.allowedCardTypes = [.release, .sdk]
         config.logConfig = Log.Config.custom(logLevel: Log.Level.allCases,
-                                             loggers: [loggerProvider.logger, ConsoleLogger()])
+                                             loggers: [FileLogger(), ConsoleLogger()])
         config.filter.batchIdFilter = .deny(["0027", // todo: tangem tags
                                              "0030",
                                              "0031",
